+++ QingToolbox.WebUI/src/design-system/components/QPage.vue
@@ -1,0 +1,1 @@
+<template><main class="q-page"><slot/></main></template>