+++ QingToolbox.WebUI/src/design-system/components/QPanel.vue
@@ -1,0 +1,1 @@
+<template><section class="q-panel"><slot/></section></template>