+++ QingToolbox.WebUI/src/design-system/components/QSkeleton.vue
@@ -1,0 +1,1 @@
+<template><div class="q-skeleton" aria-hidden="true"/></template>